Refactor: 프로젝트 구조 변경
- 클린 아키텐처 적용
