+++ app/src/main/java/com/survivalcoding/maskinfo/Config.kt
@@ -1,0 +1,3 @@
+package com.survivalcoding.maskinfo
+
+const val URL_BASE = "http://104.198.248.76:3000/"
