--- app/src/main/java/com/survivalcoding/maskinfo/Configs.kt
@@ -1,8 +1,0 @@
-package com.survivalcoding.maskinfo
-
-class Configs {
-    companion object{
-        const val URL_BASE = "http://104.198.248.76:3000/"
-        const val PERMISSION_REQUEST_CODE_LOCATION = 1000
-    }
-}
