+++ app/src/main/java/com/survivalcoding/maskinfo/domain/model/Coordinate.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.maskinfo.data.model
+package com.survivalcoding.maskinfo.domain.model
 
 data class Coordinate(
     val latitude: Double,
