+++ app/src/main/java/com/survivalcoding/maskinfo/domain/model/Info.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.maskinfo.data.model
+package com.survivalcoding.maskinfo.domain.model
 
 data class Info(
     val name: String,
