+++ app/src/main/java/com/survivalcoding/maskinfo/domain/repository/InfoRepository.kt
@@ -1,0 +1,7 @@
+package com.survivalcoding.maskinfo.domain.repository
+
+import com.survivalcoding.maskinfo.data.dto.Mask
+
+interface InfoRepository {
+    suspend fun getMask(currentPage: Int): Mask
+}
