+++ app/src/main/java/com/survivalcoding/maskinfo/presentation/main/MainViewModel.kt
@@ -1,11 +1,11 @@
-package com.survivalcoding.maskinfo.ui
+package com.survivalcoding.maskinfo.presentation.main
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.survivalcoding.maskinfo.data.model.Coordinate
-import com.survivalcoding.maskinfo.data.model.Info
-import com.survivalcoding.maskinfo.data.model.mapper.toInfo
-import com.survivalcoding.maskinfo.data.repository.InfoRepository
+import com.survivalcoding.maskinfo.domain.model.Coordinate
+import com.survivalcoding.maskinfo.domain.model.Info
+import com.survivalcoding.maskinfo.data.mapper.toInfo
+import com.survivalcoding.maskinfo.data.repository.InfoRepositoryImpl
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
@@ -13,7 +13,7 @@ import org.json.JSONException
 import java.io.IOException
 
 class MainViewModel : ViewModel() {
-    private val infoRepository: InfoRepository by lazy { InfoRepository() }
+    private val infoRepository: InfoRepositoryImpl by lazy { InfoRepositoryImpl() }
     private var _state = MutableStateFlow(MainState())
     val state = _state.asStateFlow()
     var userCoordinate: Coordinate? = null
